+++ yash-cli/tests/scripted_test/case-p.sh
@@ -189,7 +189,7 @@ case $(false) in
 esac
 __IN__
 
-# The behavior is POSIXly-unspecified for this case. See case-y.tst.
+# The behavior is POSIXly-unspecified for this case. See case-y.sh.
 #test_OE -e 0 'exit status of case command (matched, empty)'
 
 test_OE -e 17 'exit status of case command (matched, non-empty)'
@@ -199,6 +199,27 @@ case $(echo 2; exit 2) in
     3) true; (exit 19);;
 esac
 __IN__
+
+test_oE -e 42 'executing item after ;&'
+case 1 in
+    0) echo not reached 0;;
+    1) echo matched 1;&
+    2) echo matched 2; (exit 42);&
+esac
+__IN__
+matched 1
+matched 2
+__OUT__
+
+test_oE 'exit status after empty ;& in case command'
+(exit 1)
+case i in
+    i) ;&
+    j) echo $?
+esac
+__IN__
+1
+__OUT__
 
 test_oE 'patterns can be preceded by ('
 case a in